cf_v5_src/requirements.txt
# Function dependencies, for example:
# package>=version
pandas-datareader
SQLAlchemy
psycopg2
SQLAlchemy-Utils
google-cloud-pubsub
yfinance==0.2.63